Extract some code
